Move makeCtx to test-helpers, drop it from factory public API
With no production callers left, makeCtx is a test-only helper. Move it
from strategy-game-factory/helpers to a new test-helpers folder and stop
exporting it from the factory index, so prod code can no longer reach it
via the public API. Spec files import it from test-helpers directly.
