+++ src/components/games/bank-robbers/bot-strategy.spec.ts
@@ -1,6 +1,7 @@
 import { smartBotStrategy } from "./bot-strategy";
 import type { Board } from './bank-robbers';
-import { makeCtx, type GameMoves } from '../../strategy-game-factory';
+import { type GameMoves } from '../../strategy-game-factory';
+import { makeCtx } from '../../strategy-game-factory/test-helpers';
 
 const mockRob = (): GameMoves<Board> => ({
   rob: (board: Board, bank: number) => { board.circle[bank] = false; return { nextBoard: board }; }
