+++ src/components/games/dominoes-4x4/bot-strategy.spec.ts
@@ -1,4 +1,5 @@
-import { makeCtx, type GameMoves } from '../../strategy-game-factory';
+import { type GameMoves } from '../../strategy-game-factory';
+import { makeCtx } from '../../strategy-game-factory/test-helpers';
 import { smartBotStrategy, randomBotStrategy, isWinningForPlayerToMove } from './bot-strategy';
 import { getPossibleMoves, type Board, type Domino, type Field } from './dominoes-4x4';
