+++ src/components/games/number-pyramid/number-pyramid.spec.ts
@@ -1,6 +1,6 @@
 import { moves } from './number-pyramid';
 import type { Board, Slot } from './strategy';
-import { makeCtx } from '../../strategy-game-factory';
+import { makeCtx } from '../../strategy-game-factory/test-helpers';
 import { makeEvents } from '../../../test-utils';
 
 const active = (value: number): Slot => ({ value, state: 'active' });
