+++ src/components/games/polynomial-building/bot-strategy.spec.ts
@@ -1,4 +1,5 @@
-import { makeCtx, type GameMoves, type StrategyArgs } from '../../strategy-game-factory';
+import { type GameMoves, type StrategyArgs } from '../../strategy-game-factory';
+import { makeCtx } from '../../strategy-game-factory/test-helpers';
 import { smartBotStrategy, randomBotStrategy } from './bot-strategy';
 import { type Board, type Coef, hasThreeIntegerRoots, canComplete } from './helpers';
