+++ src/components/games/recolouring-discs/bot-strategy.spec.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect } from 'vitest';
-import { makeCtx, type GameMoves, type StrategyArgs } from '../../strategy-game-factory';
+import { type GameMoves, type StrategyArgs } from '../../strategy-game-factory';
+import { makeCtx } from '../../strategy-game-factory/test-helpers';
 import {
   type Board,
   type Cell,
