+++ src/components/games/tictactoe-alikes/anti-tictactoe/bot-strategy.spec.ts
@@ -1,6 +1,7 @@
 import { smartBotStrategy } from "./bot-strategy";
 import type { Board } from './helpers';
-import { makeCtx, type GameMoves } from '../../../strategy-game-factory';
+import { type GameMoves } from '../../../strategy-game-factory';
+import { makeCtx } from '../../../strategy-game-factory/test-helpers';
 
 const mockPlacePiece = (): GameMoves<Board> => ({
   placePiece: (board: Board, id: number) => { board[id] = 'new_piece'; return { nextBoard: board }; }
