+++ src/components/strategy-game-factory/test-helpers/index.ts
@@ -1,0 +1,1 @@
+export { makeCtx } from './make-ctx';
